+++ src/components/my/events/EventForm.tsx
@@ -1,59 +1,30 @@
-import { ChangeEvent, useState } from 'react';
-import { useFieldArray, useForm } from 'react-hook-form';
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
-import { TiDelete } from 'react-icons/ti';
-import DaumPostcodeEmbed from 'react-daum-postcode';
-import { CalendarIcon } from 'lucide-react';
-import { format } from 'date-fns';
-import { ko } from 'date-fns/locale';
 
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@components/ui/form';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@components/ui/select';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@components/ui/popover';
-import { Input } from '@components/ui/input';
+import { Form } from '@components/ui/form';
 import { Button } from '@components/ui/button';
-import { Calendar } from '@components/ui/calendar';
-import { RadioGroup, RadioGroupItem } from '@components/ui/radio-group';
-import { Textarea } from '@components/ui/textarea';
+import EventThumbnailInput from './input/EventThumbnailInput';
+import EventNameInput from './input/EventNameInput';
+import EventDateInput from './input/EventDateInput';
+import EventCategoryInput from './input/EventCategoryInput';
+import EventLocationInput from './input/EventLocationInput';
+import EventRegistrationDateInput from './input/EventRegistrationDateInput';
+import EventDescriptionInput from './input/EventDescriptionInput';
+import EventTicketInput from './input/EventTicketInput';
 
-import { cn } from '@/lib/utils';
+import { EventFormValues } from '@/types/Form';
 import { useGlobalAlertStore } from '@store/useGlobalAlertStore';
-import RightArrow from '@assets/images/icons/RightArrow.svg';
-import { CATEGORIES } from '@constants/categories';
 import useUser from '@hooks/useUser';
 import { createEvent } from '@services/eventService';
-import calculateEndTimeOptions from '@utils/my/calculateEndTimeOptions';
-import resizeAndConvertImage from '@utils/resizeAndConvertImage';
-import { EventFormValues } from '@/types/Form';
-
-interface PostcodeData {
-  address: string;
-  addressType: string;
-  bname: string;
-  buildingName: string;
-}
 
 function EventForm() {
   const user = useUser();
   const { openAlert } = useGlobalAlertStore();
+  const navigate = useNavigate();
+
   const form = useForm<EventFormValues>({
     mode: 'onChange',
     resolver: zodResolver(eventFormSchema),
@@ -75,11 +46,6 @@ function EventForm() {
     },
   });
 
-  const { fields, append, remove } = useFieldArray({
-    control: form.control,
-    name: 'tickets',
-  });
-
   const startDate = form.watch('startDate');
   const startTime = form.watch('startTime');
   const registrationStartDate = form.watch('registrationStartDate');
@@ -89,49 +55,6 @@ function EventForm() {
 
   const [thumbnailPreview, setThumbnailPreview] = useState<string | null>(null);
   const [isPostcodeOpen, setIsPostcodeOpen] = useState<boolean>(false);
-
-  const handleThumbnailChange = async (
-    event: ChangeEvent<HTMLInputElement>,
-  ) => {
-    const files = event.target.files;
-
-    if (files && files.length > 0) {
-      const file = files[0];
-      const webpImageBlob = await resizeAndConvertImage(file);
-
-      // Blob을 File 객체로 변환
-      const webpImageFile = new File([webpImageBlob], `${file.name}.webp`, {
-        type: 'image/webp',
-        lastModified: new Date().getTime(),
-      });
-
-      // 이미지 미리보기 설정
-      form.setValue('thumbnail', webpImageFile, { shouldValidate: true });
-      const imageUrl = URL.createObjectURL(webpImageFile);
-      setThumbnailPreview(imageUrl);
-    } else {
-      setThumbnailPreview(null);
-    }
-  };
-
-  const handlePostcodeComplete = (data: PostcodeData) => {
-    let fullAddress = data.address;
-    let extraAddress = '';
-
-    if (data.addressType === 'R') {
-      if (data.bname !== '') {
-        extraAddress += data.bname;
-      }
-      if (data.buildingName !== '') {
-        extraAddress +=
-          extraAddress !== '' ? `, ${data.buildingName}` : data.buildingName;
-      }
-      fullAddress += extraAddress !== '' ? ` (${extraAddress})` : '';
-    }
-
-    form.setValue('location', fullAddress);
-    setIsPostcodeOpen(false);
-  };
 
   const togglePostcode = () => {
     setIsPostcodeOpen(prev => !prev);
@@ -147,6 +70,7 @@ function EventForm() {
       .then(result => {
         if (result.success) {
           openAlert('이벤트가 등록되었습니다.', '');
+          navigate('/my/events');
         } else {
           openAlert('다시 시도해 주세요.', result.error as string);
         }
@@ -160,620 +84,41 @@ function EventForm() {
     <div>
       <Form {...form}>
         <form onSubmit={form.handleSubmit(onSubmit)} className='space-y-8'>
-          <div className='flex flex-col space-y-2'>
-            <p className='py-1 text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-              썸네일
-            </p>
-            {thumbnailPreview ? (
-              <div className='flex h-[320px] w-full items-center justify-center overflow-hidden rounded-md border border-input'>
-                <img
-                  src={thumbnailPreview}
-                  alt='Thumbnail Preview'
-                  className='size-full h-full w-full object-cover'
-                />
-              </div>
-            ) : (
-              <div className='flex h-[320px] w-full items-center justify-center overflow-hidden rounded-md border border-input'>
-                <p className='text-center text-sm text-gray-500'>
-                  썸네일 미리보기
-                </p>
-              </div>
-            )}
-            <div className='flex justify-end'>
-              <label
-                htmlFor='picture'
-                className='inline-flex h-fit cursor-pointer items-center justify-center whitespace-nowrap rounded-md border border-input bg-background px-4 py-2 text-xs font-medium ring-offset-background transition-colors hover:bg-accent hover:text-accent-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50'
-              >
-                업로드
-              </label>
-            </div>
-            <FormField
-              name='thumbnail'
-              render={() => (
-                <FormItem>
-                  <FormControl>
-                    <Input
-                      id='picture'
-                      type='file'
-                      accept='image/*'
-                      className='hidden'
-                      {...form.register('thumbnail')}
-                      onChange={handleThumbnailChange}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-          </div>
-
-          <div className='flex flex-col space-y-2'>
-            <p className='py-1 text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-              이벤트 이름
-            </p>
-            <FormField
-              name='name'
-              render={() => (
-                <FormItem>
-                  <FormControl>
-                    <Input
-                      placeholder='이벤트 이름을 입력해 주세요.'
-                      {...form.register('name', {
-                        setValueAs: value => value.trim(),
-                      })}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-          </div>
-
-          <div className='flex flex-col space-y-2'>
-            <p className='py-1 text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-              이벤트 일시
-            </p>
-            <div className='flex flex-wrap'>
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  name='startDate'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Popover>
-                        <PopoverTrigger asChild>
-                          <FormControl>
-                            <Button
-                              variant={'outline'}
-                              className={cn(
-                                'w-[240px] pl-3 text-left font-normal',
-                                !field.value && 'text-muted-foreground',
-                              )}
-                            >
-                              {field.value ? (
-                                format(field.value, 'PPP', { locale: ko })
-                              ) : (
-                                <span>시작 날짜를 선택해 주세요.</span>
-                              )}
-                              <CalendarIcon className='ml-auto h-4 w-4 opacity-50' />
-                            </Button>
-                          </FormControl>
-                        </PopoverTrigger>
-                        <PopoverContent className='w-auto p-0' align='start'>
-                          <Calendar
-                            mode='single'
-                            selected={
-                              field.value ? new Date(field.value) : undefined
-                            }
-                            onSelect={field.onChange}
-                            disabled={date => {
-                              const registrationEndDateValue =
-                                registrationEndDate
-                                  ? new Date(registrationEndDate)
-                                  : null;
-                              return registrationEndDateValue
-                                ? date <= registrationEndDateValue
-                                : new Date() > date;
-                            }}
-                            initialFocus
-                          />
-                        </PopoverContent>
-                      </Popover>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  rules={{ required: '시작 시간은 필수입니다.' }}
-                  name='startTime'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Select
-                        onValueChange={field.onChange}
-                        defaultValue={field.value}
-                      >
-                        <FormControl>
-                          <SelectTrigger>
-                            <SelectValue placeholder='시작 시간을 선택해 주세요.' />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {Array.from({ length: 48 }, (_, index) => {
-                            const hour = Math.floor(index / 2);
-                            const minute = (index % 2) * 30;
-                            const timeString = `${hour
-                              .toString()
-                              .padStart(2, '0')}:${minute
-                              .toString()
-                              .padStart(2, '0')}`;
-                            const displayHour =
-                              hour % 12 === 0 ? 12 : hour % 12;
-                            const ampm = hour < 12 ? '오전' : '오후';
-                            return (
-                              <SelectItem key={timeString} value={timeString}>
-                                {`${ampm} ${displayHour}시 ${minute
-                                  .toString()
-                                  .padStart(2, '0')}분`}
-                              </SelectItem>
-                            );
-                          })}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-
-              <div className='flex flex-grow basis-[5%] items-center justify-center'>
-                <img src={RightArrow} alt='RightArrow' />
-              </div>
-
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  name='endDate'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Popover>
-                        <PopoverTrigger asChild>
-                          <FormControl>
-                            <Button
-                              variant={'outline'}
-                              className={cn(
-                                'w-[240px] pl-3 text-left font-normal',
-                                !field.value && 'text-muted-foreground',
-                              )}
-                            >
-                              {field.value ? (
-                                format(field.value, 'PPP', { locale: ko })
-                              ) : (
-                                <span>종료 날짜를 선택해 주세요.</span>
-                              )}
-                              <CalendarIcon className='ml-auto h-4 w-4 opacity-50' />
-                            </Button>
-                          </FormControl>
-                        </PopoverTrigger>
-                        <PopoverContent className='w-auto p-0' align='start'>
-                          <Calendar
-                            mode='single'
-                            selected={
-                              field.value ? new Date(field.value) : undefined
-                            }
-                            onSelect={field.onChange}
-                            disabled={date =>
-                              startDate
-                                ? date < new Date(startDate)
-                                : date < new Date()
-                            }
-                            initialFocus
-                          />
-                        </PopoverContent>
-                      </Popover>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  rules={{ required: '종료 시간은 필수입니다.' }}
-                  name='endTime'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Select
-                        onValueChange={field.onChange}
-                        defaultValue={field.value}
-                        disabled={!startDate || !startTime}
-                      >
-                        <FormControl>
-                          <SelectTrigger>
-                            <SelectValue placeholder='종료 시간을 선택해 주세요.' />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {startDate &&
-                            calculateEndTimeOptions(startDate, startTime).map(
-                              option => (
-                                <SelectItem
-                                  key={option.value}
-                                  value={option.value}
-                                >
-                                  {option.label}
-                                </SelectItem>
-                              ),
-                            )}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-            </div>
-          </div>
-
-          <FormField
-            control={form.control}
-            name='category'
-            render={({ field }) => (
-              <FormItem className='space-y-3'>
-                <FormLabel>카테고리</FormLabel>
-                <FormControl>
-                  <RadioGroup
-                    onValueChange={field.onChange}
-                    defaultValue={field.value}
-                    className='flex flex-wrap gap-x-0 gap-y-3'
-                  >
-                    {Object.entries(CATEGORIES).map(([key, label]) => (
-                      <FormItem
-                        key={key}
-                        className='flex w-1/5 items-center space-x-2'
-                      >
-                        <FormControl>
-                          <RadioGroupItem value={key} />
-                        </FormControl>
-                        <FormLabel className='!my-0 font-normal'>
-                          {label}
-                        </FormLabel>
-                      </FormItem>
-                    ))}
-                  </RadioGroup>
-                </FormControl>
-              </FormItem>
-            )}
+          <EventThumbnailInput
+            form={form}
+            thumbnailPreview={thumbnailPreview}
+            setThumbnailPreview={setThumbnailPreview}
           />
 
-          <div className='flex flex-col space-y-2'>
-            <Input
-              type='hidden'
-              {...form.register('location', {
-                required: '이벤트 장소 입력은 필수입니다.',
-              })}
-              value={location}
-              onChange={({ target }) => form.setValue('location', target.value)}
-            />
-            <p className='py-1 text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-              이벤트 장소
-            </p>
-            <div
-              className='w-full cursor-pointer overflow-hidden rounded-md border border-input px-3 py-2 text-sm'
-              onClick={togglePostcode}
-            >
-              {location ? location : '주소를 검색해 주세요.'}
-            </div>
-            {form.formState.errors.location && (
-              <p className='text-sm font-medium text-destructive'>
-                {form.formState.errors.location.message}
-              </p>
-            )}
-            {isPostcodeOpen && (
-              <div>
-                <DaumPostcodeEmbed
-                  onComplete={handlePostcodeComplete}
-                  autoClose
-                  className='z-10 h-[450px] w-full'
-                />
-              </div>
-            )}
-          </div>
+          <EventNameInput form={form} />
 
-          <div className='flex flex-col space-y-2'>
-            <p className='py-1 text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-              모집 기간{' '}
-              <span className='text-xs font-normal text-gray-500'>
-                (이벤트 일시를 먼저 선택해 주세요.)
-              </span>
-            </p>
-            <div className='flex flex-wrap'>
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  name='registrationStartDate'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Popover>
-                        <PopoverTrigger asChild>
-                          <FormControl>
-                            <Button
-                              variant={'outline'}
-                              className={cn(
-                                'w-[240px] pl-3 text-left font-normal',
-                                !field.value && 'text-muted-foreground',
-                              )}
-                            >
-                              {field.value ? (
-                                format(field.value, 'PPP', { locale: ko })
-                              ) : (
-                                <span>시작 날짜를 선택해 주세요.</span>
-                              )}
-                              <CalendarIcon className='ml-auto h-4 w-4 opacity-50' />
-                            </Button>
-                          </FormControl>
-                        </PopoverTrigger>
-                        <PopoverContent className='w-auto p-0' align='start'>
-                          <Calendar
-                            mode='single'
-                            selected={
-                              field.value ? new Date(field.value) : undefined
-                            }
-                            onSelect={field.onChange}
-                            disabled={date => {
-                              const startDateValue = startDate
-                                ? new Date(startDate)
-                                : new Date();
-                              return (
-                                date < new Date() || date >= startDateValue
-                              );
-                            }}
-                            initialFocus
-                          />
-                        </PopoverContent>
-                      </Popover>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
+          <EventDateInput
+            form={form}
+            startDate={startDate}
+            startTime={startTime}
+            registrationEndDate={registrationEndDate}
+          />
 
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  rules={{ required: '시작 시간은 필수입니다.' }}
-                  name='registrationStartTime'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Select
-                        onValueChange={field.onChange}
-                        defaultValue={field.value}
-                      >
-                        <FormControl>
-                          <SelectTrigger>
-                            <SelectValue placeholder='시작 시간을 선택해 주세요.' />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {Array.from({ length: 48 }, (_, index) => {
-                            const hour = Math.floor(index / 2);
-                            const minute = (index % 2) * 30;
-                            const timeString = `${hour
-                              .toString()
-                              .padStart(2, '0')}:${minute
-                              .toString()
-                              .padStart(2, '0')}`;
-                            const displayHour =
-                              hour % 12 === 0 ? 12 : hour % 12;
-                            const ampm = hour < 12 ? '오전' : '오후';
-                            return (
-                              <SelectItem key={timeString} value={timeString}>
-                                {`${ampm} ${displayHour}시 ${minute
-                                  .toString()
-                                  .padStart(2, '0')}분`}
-                              </SelectItem>
-                            );
-                          })}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
+          <EventCategoryInput form={form} />
 
-              <div className='flex flex-grow basis-[5%] items-center justify-center'>
-                <img src={RightArrow} alt='RightArrow' />
-              </div>
+          <EventLocationInput
+            form={form}
+            location={location}
+            isPostcodeOpen={isPostcodeOpen}
+            togglePostcode={togglePostcode}
+            setIsPostcodeOpen={setIsPostcodeOpen}
+          />
 
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  name='registrationEndDate'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Popover>
-                        <PopoverTrigger asChild>
-                          <FormControl>
-                            <Button
-                              variant={'outline'}
-                              className={cn(
-                                'w-[240px] pl-3 text-left font-normal',
-                                !field.value && 'text-muted-foreground',
-                              )}
-                            >
-                              {field.value ? (
-                                format(field.value, 'PPP', { locale: ko })
-                              ) : (
-                                <span>종료 날짜를 선택해 주세요.</span>
-                              )}
-                              <CalendarIcon className='ml-auto h-4 w-4 opacity-50' />
-                            </Button>
-                          </FormControl>
-                        </PopoverTrigger>
-                        <PopoverContent className='w-auto p-0' align='start'>
-                          <Calendar
-                            mode='single'
-                            selected={
-                              field.value ? new Date(field.value) : undefined
-                            }
-                            onSelect={field.onChange}
-                            disabled={date => {
-                              const registrationStartDateValue =
-                                registrationStartDate
-                                  ? new Date(registrationStartDate)
-                                  : new Date();
-                              const eventStartDateValue = startDate
-                                ? new Date(startDate)
-                                : new Date();
-                              return (
-                                date < registrationStartDateValue ||
-                                date >= eventStartDateValue
-                              ); // 모집 시작 날짜 이후이면서 이벤트 시작 날짜 이전 날짜만 선택 가능
-                            }}
-                            initialFocus
-                          />
-                        </PopoverContent>
-                      </Popover>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
+          <EventRegistrationDateInput
+            form={form}
+            startDate={startDate}
+            registrationStartDate={registrationStartDate}
+            registrationStartTime={registrationStartTime}
+          />
 
-              <div className='flex-grow basis-[22%]'>
-                <FormField
-                  control={form.control}
-                  rules={{ required: '종료 시간은 필수입니다.' }}
-                  name='registrationEndTime'
-                  render={({ field }) => (
-                    <FormItem className='flex flex-grow flex-col'>
-                      <Select
-                        onValueChange={field.onChange}
-                        defaultValue={field.value}
-                        disabled={
-                          !registrationStartDate || !registrationStartTime
-                        }
-                      >
-                        <FormControl>
-                          <SelectTrigger>
-                            <SelectValue placeholder='종료 시간을 선택해 주세요.' />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {registrationStartDate &&
-                            registrationStartTime &&
-                            calculateEndTimeOptions(
-                              registrationStartDate,
-                              registrationStartTime,
-                            ).map(option => (
-                              <SelectItem
-                                key={option.value}
-                                value={option.value}
-                              >
-                                {option.label}
-                              </SelectItem>
-                            ))}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-            </div>
-          </div>
+          <EventDescriptionInput form={form} />
 
-          <div className='flex flex-col space-y-2'>
-            <p className='py-1 text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-              이벤트 이름
-            </p>
-            <FormField
-              control={form.control}
-              name='description'
-              render={({ field }) => (
-                <FormItem>
-                  <FormControl>
-                    <Textarea
-                      placeholder='이벤트 소개 내용을 마크다운 형식으로 입력해 주세요.'
-                      className='min-h-60 resize-none'
-                      {...field}
-                    />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-          </div>
-
-          <div className='flex flex-col gap-2 '>
-            {fields.map((field, index) => (
-              <div key={field.id} className='flex gap-2'>
-                <div className='flex basis-[32%] items-center gap-2'>
-                  <p className='text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-                    티켓 이름
-                  </p>
-                  <Input
-                    {...form.register(`tickets.${index}.name`)}
-                    placeholder='일반 티켓'
-                    className='w-[70%]'
-                  />
-                </div>
-                <div className='flex basis-[32%] items-center gap-2'>
-                  <p className='text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-                    티켓 가격
-                  </p>
-                  <Input
-                    {...form.register(`tickets.${index}.price`, {
-                      setValueAs: value => parseInt(value, 10) || 0,
-                    })}
-                    placeholder='20000'
-                    type='number'
-                    className='w-[70%]'
-                  />
-                </div>
-                <div className='flex basis-[32%] items-center gap-2'>
-                  <p className='text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'>
-                    판매 개수
-                  </p>
-                  <Input
-                    {...form.register(`tickets.${index}.quantity`, {
-                      setValueAs: value => parseInt(value, 10) || 0,
-                    })}
-                    placeholder='1000'
-                    type='number'
-                    className='w-[70%]'
-                  />
-                </div>
-                <div className='flex basis-[5%] justify-end'>
-                  <button type='button' onClick={() => remove(index)}>
-                    <TiDelete size='20' />
-                  </button>
-                </div>
-              </div>
-            ))}
-            <div>
-              {form.formState.errors.tickets?.root && (
-                <p className='text-sm text-red-500'>
-                  {form.formState.errors.tickets.root.message}
-                </p>
-              )}
-            </div>
-            <div className='flex justify-end'>
-              <Button
-                type='button'
-                variant='outline'
-                onClick={() => append({ name: '', price: 0, quantity: 0 })}
-                className='h-fit text-xs'
-              >
-                티켓 추가
-              </Button>
-            </div>
-          </div>
+          <EventTicketInput form={form} />
 
           <Button className='!mt-16 w-full'>등록하기</Button>
         </form>
@@ -792,7 +137,6 @@ const ticketSchema = z.object({
 
 const eventFormSchema = z.object({
   thumbnail: z.any().refine(val => val instanceof Blob, {
-    // Blob 체크 (File은 Blob의 확장이므로 이 조건을 만족합니다)
     message: '썸네일은 필수입니다.',
   }),
   name: z.string().min(1, '이벤트 이름은 필수입니다.').trim(),
